backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gspread==6.1.3
python-dotenv==1.0.1
pydantic==2.9.2
